pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Radjin Sardjoe Missier <rmissier@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.80.0"
uvicorn = "^0.18.3"
tortoise-orm = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.6.0"
pylint = "^2.14.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
